ICU-535 casted HP/UX warning on size_t to int
X-SVN-Rev: 3726

+++ icu4c/source/i18n/rbbi.cpp
@@ -593,18 +593,18 @@ BreakIterator *  RuleBasedBreakIterator::createBufferClone(void *stackBuffer,
     }
     if (text == NULL)
     {
-        bufferSizeNeeded = sizeof(RuleBasedBreakIterator);
+        bufferSizeNeeded = (int32_t) sizeof(RuleBasedBreakIterator);
         IterIsString = IterIsUChar = FALSE;
     }
     else if (text->getDynamicClassID() == StringCharacterIterator::getStaticClassID()) 
     {
-        bufferSizeNeeded = sizeof(struct bufferCloneStructString);
+        bufferSizeNeeded = (int32_t) sizeof(struct bufferCloneStructString);
         IterIsString = TRUE;
         IterIsUChar = FALSE;
     } 
     else if (text->getDynamicClassID() == UCharCharacterIterator::getStaticClassID()) 
     {
-        bufferSizeNeeded = sizeof(struct bufferCloneStructUChar);
+        bufferSizeNeeded = (int32_t) sizeof(struct bufferCloneStructUChar);
         IterIsString = FALSE;
         IterIsUChar = TRUE;
     }
